media: ti: j721e-csi2rx: Fix locking in restart_dma
After the frame is submitted to DMA, it may happen that the submitted
list is not updated soon enough, and the DMA callback is triggered
before that. This can lead to kernel crashes, so move everything in a
single lock/unlock section to prevent such races.

Fixes: b93eb4cf5ccb ("media: ti: j721e-csi2rx: Submit all available buffers")
Signed-off-by: Jai Luthra <j-luthra@ti.com>
Reviewed-by: Kamlesh Gurudasani <kamlesh@ti.com>
(cherry picked from commit d23b3d303a07057885900044db53102bd746703a)