net: mdiobus: Add support for optional PHY regulator control
Some hardware platforms support disabling power to the ethernet phy
to reduce power consumption.

This patch adds support for managing an optional regulator for PHY
devices on the MDIO bus. It introduces a new function,
mdiobus_register_regulator, which attempts to obtain a regulator
named "phy-supply" for the MDIO device. If the regulator is present and
successfully obtained, it will be enabled during PHY reset deassertion
and disabled during PHY reset assertion.

Signed-off-by: Nate Drude <nate.d@variscite.com>